services/agent-gateway/pyproject.toml
[project]
name = "agent-gateway"
version = "0.1.0"
description = "WebSocket gateway for AI agent connections to Convene AI meetings"
requires-python = ">=3.12"
dependencies = [
    "convene-core",
    "convene-providers",
    "audio-service",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13.0",
    "redis>=5.0",
    "pydantic-settings>=2.0",
    "PyJWT>=2.8",
]

[tool.uv.sources]
convene-core = { workspace = true }
convene-providers = { workspace = true }
audio-service = { workspace = true }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/agent_gateway"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
